Move textproto to internal/textproto
