+++ builder.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/jhillyerd/enmime/internal/stringutil"
-	"github.com/jhillyerd/enmime/textproto"
+	"github.com/jhillyerd/enmime/internal/textproto"
 )
 
 // MailBuilder facilitates the easy construction of a MIME message.  Each manipulation method
