+++ dsn/dsn.go
@@ -4,8 +4,6 @@ package dsn
 
 import (
 	"strings"
-
-	"github.com/jhillyerd/enmime/textproto"
 )
 
 // Report represents delivery status report as per https://datatracker.ietf.org/doc/html/rfc6522.
