+++ dsn/parse.go
@@ -8,7 +8,6 @@ import (
 	"io"
 
 	"github.com/jhillyerd/enmime"
-	"github.com/jhillyerd/enmime/textproto"
 )
 
 // ParseReport parses p as a "container" for delivery status report (per rfc6522) if p is "multipart/report".
