+++ dsn/parse_internal_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/jhillyerd/enmime"
-	"github.com/jhillyerd/enmime/textproto"
+
 	"github.com/stretchr/testify/assert"
 )
