+++ encode.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/jhillyerd/enmime/internal/coding"
 	"github.com/jhillyerd/enmime/internal/stringutil"
-	"github.com/jhillyerd/enmime/textproto"
+	"github.com/jhillyerd/enmime/internal/textproto"
 )
 
 // b64Percent determines the percent of non-ASCII characters enmime will tolerate before switching
