+++ internal/textproto/pipeline.go
