+++ internal/textproto/reader.go
