+++ internal/textproto/reader_email.go
