+++ internal/textproto/reader_test.go
