+++ internal/textproto/writer.go
