+++ internal/textproto/writer_test.go
