+++ part.go
@@ -14,8 +14,9 @@ import (
 
 	"github.com/gogs/chardet"
 	"github.com/jhillyerd/enmime/internal/coding"
+	"github.com/jhillyerd/enmime/internal/textproto"
 	"github.com/jhillyerd/enmime/mediatype"
-	"github.com/jhillyerd/enmime/textproto"
+
 	"github.com/pkg/errors"
 )
